page

+++ src/Pages/Departments/Department.jsx
@@ -38,34 +38,42 @@ function Department() {
     // const setId = (id) => {
     //     console.log(id)
     // }
-    return (
+    const role = localStorage.getItem('role')
+    if (role === "manager") {
+        return (
 
-        <div className=''>
+            <div className=''>
 
-            <h1>Department page</h1>
+                <h1>Department page</h1>
 
-            <div className='department'>
+                <div className='department'>
 
-                <div className='btn'>
-                    <Button>
-                        <NavLink to='/addDepartment'>
-                            new add
-                        </NavLink>
-                    </Button>
-                </div>
-                {name && name.length ? '' : 'No Item...'}
-                {name.map((department) => (
-                    <div className='icons' >
-                        <strong key={department.id}>{department.departmentName}</strong>
-                        <div className='icon-cate'>
-                            <NavLink to={`/editDepartment/${department.id}`} onChange={() => setToLocalStorage(department.id, department.departmentName)}><span title="edit"><Button><FaEdit /></Button></span></NavLink>
-                            <Button><span title="delete" onClick={() => deleteItem(department.id)}><FaTrash /></span></Button>
-                        </div>
+                    <div className='btn'>
+                        <Button>
+                            <NavLink to='/addDepartment'>
+                                new add
+                            </NavLink>
+                        </Button>
                     </div>
-                ))}
+                    {name && name.length ? '' : 'No Item...'}
+                    {name.map((department) => (
+                        <div className='icons' >
+                            <strong key={department.id}>{department.departmentName}</strong>
+                            <div className='icon-cate'>
+                                <NavLink to={`/editDepartment/${department.id}`} onChange={() => setToLocalStorage(department.id, department.departmentName)}><span title="edit"><Button><FaEdit /></Button></span></NavLink>
+                                <Button><span title="delete" onClick={() => deleteItem(department.id)}><FaTrash /></span></Button>
+                            </div>
+                        </div>
+                    ))}
+                </div>
             </div>
-        </div>
-    );
+        );
+    } else {
+        return (
+
+            <h1> no Auth</h1>
+        )
+    }
 }
 
 export default Department;
